collectd: Only get interface metrics for lo0/eth0
VMs that host containers, such as Concourse and Diego cells, create and
destroy virtual network interfaces very frequently. With collectd's default
configuration of the `interface` plugin this results in Graphite/Carbon
allocating storage for a new Whisper file every time a container is started.

These metrics aren't useful because we can't tell what container each
randomly-generated interface name belonged to. They consume a lot of space
on disk, even with automatic pruning of stale metrics after 10 days. Current
usage on prod:

    bosh_zczgqkgrr@91b8d5a4-55ee-4a6e-ac9f-27f6a2b945f9:~$ du -csh /var/vcap/store/graphite/storage/whisper/collectd/{concourse,cell}_*/interface
    61G     /var/vcap/store/graphite/storage/whisper/collectd/concourse_collectd_0/interface
    10G     /var/vcap/store/graphite/storage/whisper/collectd/cell_0/interface
    40G     /var/vcap/store/graphite/storage/whisper/collectd/cell_1/interface
    14G     /var/vcap/store/graphite/storage/whisper/collectd/cell_2/interface
    29G     /var/vcap/store/graphite/storage/whisper/collectd/cell_3/interface
    153G    total

To stop us collecting these I am restricting collectd to only get metrics
for interfaces that relate to the host; `lo0` and `eth0`. I'm explicitly
setting `IgnoreSelected` because the documentation isn't very clear that
`false` is the default value.

This configuration *might* result in no network interface metrics being
collected if we upgrade to a stemcell with a later Ubuntu release because of
changes to udev device naming. I can't see a nice way to guard against it
though:

- https://bugs.launchpad.net/ubuntu/+source/ubuntu-meta/+bug/1347859

This change will deploy automatically to BOSH and CF, although it will take
some time to re-provision BOSH with `bosh-init`. It will need to be manually
deployed to the Deployer Concourse by re-running the `create-deployer`
pipeline. The stale metrics will be removed after 10 days but we might need
to intervene if `graphite/0` runs out of disk space in the meantime.