Create bucket to store ELB access logs
Create a bucket called ${env}-elb-access-log to store the access
log from different ELB.

We need to configure a associated policy to allow the ELBs to upload
the logs to the bucket, as described in the documentation[1]

In the policy we must specify the right AWS ELB account ID, listed in
the documentation. We save this as a map in the globals.tf, to make
it more explicit, allow easy update and reuse.

We do not specify the ELB prefix in the ARN of the resource of the policy,
instead we use *, so that we can reuse the same bucket for different
ELBs.

We do not specify the account ID in the ARN of the resource of the  policy.
Instead we use *, to avoid have to hardcode and manage such value across
accounts.

We set a global lifecycle policy of 30 days retention[2] for all the
objects in the bucket (prefix=""). This matches the default retention
of our logs in logstash.

Note: in aws_s3_bucket.lifecycle_rule.prefix must be defined despite
what the documentation says, or you will get the error:

  aws_s3_bucket.elb_access_log: "lifecycle_rule.0.prefix": required field is not set

[1] http://docs.aws.amazon.com/ElasticLoadBalancing/latest/DeveloperGuide/enable-access-logs.html
[2] http://docs.aws.amazon.com/AmazonS3/latest/dev/intro-lifecycle-rules.html