Remove bosh reference manifest test
We only do this for BOSH, not CF or Concourse. I believe it was
originally introduced to catch mistakes when refactoring either from
Alpha to Beta or from spiff to spruce.
I think the reference comparison test is now slowing us down rather than
helping us. Each time we make a change to the BOSH manifests we have to
remember to copy/paste the same change to the reference manifest. We
have more detailed tests that make assertions about specific properties
in the rendered manifest. So I think we should remove this test.