Add prefix for ELB access log bucket
S3 bucket names are global and shared across all AWS users. This can
cause trouble, as the name for some buckets might have been already
used by other users. This was the case for `prod-elb-access-log`.

To avoid this, we add a prefix to the ELB access log bucket to
"ensure" uniqueness and avoid collision with other 3rd parties names.

We create a variable `assets_prefix` to be used in any other place.