Avoid unnecessary godep restore in platform-tests
The availability tests have vendored dependenvies, so there's no need to
run `godep restore`. Not doing this will reduce the startup time for
this suite.

`godep restore` was previously run in the global `run_tests.sh` script.
This moves it into the run scripts for the individual suites that need
it.