Fix units and null display on app requests graph
When there are no requests, the router doesn't submit stats, resulting
in nulls in graphite. Configuring this graph to display nulls as zero
makes the graph look correct.

This also updates the graph to display requests/sec, as that's a more
usual unit for this sort of graph. As an aside, we were previously
displaying requests/min by multiplying the values by 60, which wasn't
entirely correct.