Fix units and null display on API requests graph
Similarly to the app requests graph, we want this displayed in
requests/sec.

When there are no requests, there are no stats reported, resulting in
nulls in graphite. Configuring this graph to display nulls as zero makes
the graph look correct.

This also updates the metrics to do the derivative function first. This
avoids any sudden jomps in the graph that can be caused by jumps/gaps in
the data being masked by the sum function. See
http://www.jilles.net/perma/2013/08/22/how-to-do-graphite-derivatives-correctly/
for more details.