Use HTTP healthcheck for UAA ELB
UAA ELB is configured as a HTTPS-HTTPS ELB, but the healthcheck is
simply checking the TCP port 8080. Using TCP is not really checking
the service and the healthcheck access is not being logged by the UAA
service. We were having some issues in the ELB with instances being
mark as unhealthy, but there are no logs of the healthchecks.

UAA provides an endpoint `/healthz`[1], that is being used by
monit[2][3] and which we can use from the ELB.

Note that although monit uses 8989, that is only accessible locally,
to access from the ELB we must use the public port 8080.

[1] https://github.com/cloudfoundry/uaa/blob/616e8bcc58f9bc16f9d9ec806da1b0bf0f5dae85/server/src/main/java/org/cloudfoundry/identity/uaa/health/HealthzEndpoint.java#L30
[2] https://github.com/cloudfoundry/uaa-release/blob/9946ec1323b8175a0ea3e8dc68286441e5fc682e/jobs/uaa/templates/health_check
[3] https://github.com/cloudfoundry/uaa-release/blob/9946ec1323b8175a0ea3e8dc68286441e5fc682e/jobs/uaa/monit#L7