Use govuk-lint to lint our code instead of rubocop
We want adopt the govuk-lint gem in our ruby build process, since
govuk have already defined rules for a well-formatted ruby code.

This gem is based on rubocop and we can replace it in our dependencies.

The govuk-lint-ruby command does not support the -c/--config option, but
will parse a file called .rubocop.yml[1]

We must use govuk-lint >1.2.1, which allows to override the
`AllCops/Exclude` configuration [2][3]

[1] https://github.com/alphagov/govuk-lint/pull/17
[2] https://github.com/alphagov/govuk-lint/pull/56
[3] https://github.com/alphagov/govuk-lint/pull/57