govuk-lint: fix Style/ClassVars
Fix offence:

   Style/ClassVars: Replace class var @@manifest_with_defaults with a class instance var. [1]

In this case we want to use a singleton because we don't want to parse
the manifests and fixtures in each test, which would be really slow
(80s vs 3s).

But to workaround the issue, we use the Singleton module[2] in a Cache
class inside the Helpers module.

[1] https://github.com/bbatsov/ruby-style-guide#no-class-vars
[2] http://ruby-doc.org/stdlib-1.9.3/libdoc/singleton/rdoc/Singleton.html