Retitle CPU graphs
In d64ddcf and d135733 we switched to storing rates of jiffies and displaying
in grafana as the sum of the percentage use across all cores. For example, on
a 2 core machine the usage will be between 0 and 200%.

This commit tries to make this clear in the VM resource usage dashboard CPU
graph.