Identify and remove surplus testing users
Here we extend common delete_admin task to remove any users that match
the same prefix created in the create_admin task.

This is implemented in terms of `cf curl` to query the list all
users api[1]() api, and use of `jq` to identify those matching the
PREFIX that create_admin creates.

As a side effect the `delete_admin` task now needs the same PREFIX
parameter as the `create_admin` task.

As a step of extra checking, the delete_admin scriptlet now runs with -u
to guard against undefined values leading to a deletion of all users
starting with an undefined prefix.

This commit is intended to be short-lived and should be reverted once
cleanup has propogated through production environments.