Make errors from pipecleaner more explicit
Currently the script to lint concourse pipelines, pipecleaner.py,
uses a color code to report the warnings (yellow) or errors (red).

This is not so intuitive, and also in the travis output there are no
colors, making it difficult to identify the root cause of the
failure.[1]

We add a text "WARNING" or "ERROR" in each case, keeping the colors, to
make it more explicit.

[1] https://travis-ci.org/alphagov/paas-cf/builds/152994177#L751