Use `cf remove-user` to remove user from test jobs
Here we rework the common delete_admin.yml task to use cf(1) rather
than uaac(1) when deleting the user, so that the deletion is progated to
the cloudcontroller.

There is still an asymetry here in that we use uaac to make the testing
admin user, as you cannot create an admin with `cf create-user`.