Fix error in `queue` VM disk type declaration
During story #119269637[1] this typo was introduced, which led to
Bosh silently ignoring the `disk_type` field and applying only the
ephemeral disk defined in the `small` `vm_type` declaration.

The field name has been corrected. This will cause Bosh to attach
a persistent disk to the VM alongside the existing ephemeral disk.

[1] https://www.pivotaltracker.com/story/show/119269637