Add chrome_plugin to filter Pingdom status checks
This is a small chrome plugin filters undesired checks.
It workarounds the issue of using a shared Pingdom Account with checks
from multiple teams: Pingdom only provides one status page per account,
and we want to display only our checks in the dashboard.

It keeps only the checks which name starts with "PaaS"

It also strips away some of the cruft:
* The header a footer
* The page selector
* The page title

The script is also implemented as a userscript.