Move consul and etcd VMs to gp2 persistent disk
We want the etcd and consul VMs to use gp2 solid state drives, instead
of the standard drives, to achieve greater I/O performance[1] and to
be more consistent with the other VMs in our deployment.

Create disk type definitions and reference them in the jobs. This
will cause Bosh to migrate the persistent disks to gp2.

[1] http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ebs-io-characteristics.html