Cleanup users with no username
Users with no name occur in the cloudcontroller database when the user
is deleted with uaac outside of the view of the cloudcontroller.

Here we add a task to `cf-deploy` job called `purge-unnamed-users`
to query all the users out of the CC API that have no username, and
then request an asynchronous deletion of them.

This housekeeping may not be required in the longterm, in which case
this commit should be safe to revert in isolation.