Increase timeout for tag_release tests
Similar to a80adef35eb68606abfefb29ad58471f9d1c33e7

This script sometimes takes more than Eventually's default timeout of 1sec to
return when running in travis. Probably git operations are too heavy for
the container based travis build.

This happened in travis several times[1]. Dan was able to reproduce it
by running `ginkgo --untilItFails`, but I couldn't.

I set the timeout to 3 as I don't mind wait one second extra in the rare
case it timeouts and it is a nice number.

The failure looked like this:

  • Failure in Spec Setup (BeforeEach) [2.847 seconds]
  TagRelease
  /home/travis/build/alphagov/paas-cf/concourse/scripts/tag_release_test.go:147
    when there are multiple 'previous-*' tags to be promoted
    /home/travis/build/alphagov/paas-cf/concourse/scripts/tag_release_test.go:146
      and the before last tag of 'previous-*' is promoted [BeforeEach]
      /home/travis/build/alphagov/paas-cf/concourse/scripts/tag_release_test.go:115
        should promote the same tag version for 'next-*'
        /home/travis/build/alphagov/paas-cf/concourse/scripts/tag_release_test.go:89
        Timed out after 1.000s.
        Expected process to exit.  It did not.
        /home/travis/build/alphagov/paas-cf/concourse/scripts/tag_release_test.go:161

[1] https://travis-ci.org/alphagov/paas-cf/builds/153593309