Automate rotation of Cloud Foundry user password
We want an automated, repeatable, and standardised way of rotating
a Cloud Foundry user's password and emailing it to them when they
are unable to log in.

Common functions have been taken from the `create-user.sh` script,
so they have been split into a separate file. This means Shellcheck
is now configured to check included source files.

Run the script without arguments to see its documentation.