Move deployment key generation to deployer
Because we have now separate keys for concourse
(https://github.com/alphagov/paas-cf/pull/437) and BOSH
(https://github.com/alphagov/paas-cf/pull/441) we can move
"deployment ssh key" generation to create-bosh-cf pipeline.

This is the key that BOSH uses for all the VMs it deploys. The
key has thus be already available for BOSH creation, therefore
we move it from VPC terraform (and outputs) to BOSH terraform.
We generate it at the same time we generate other deployment
credentials.