Revert "Cleanup users with no username"
This reverts commit c39578ee7261336fb0f5b30c59af46bcf4e8b7fd.

The script is no longer needed now that: it has run in all persistent
environments (including production) and we have changed the tests in 7b63a98
to no longer leave these users behind.

Furthermore we had some concerns about the safety of running this long term,
as it is deleting users and we don't have any tests around it.