secret_generator: s/ignores/removes/ in test desc
Change the test description to make it clearer that existing passwords which
are no longer "requested" will be removed from the output. This is based on
a discussion with Alex when reviewing #442, where I asked if we needed to do
this manually and he pointed out that the code already does it.