manually_upload_certs: Use Terraform remote state
This makes it consistent with the `set-up-pingdom.sh` script which was
modified in b01f941 to use Terraform's remote state functionality. This is
simpler and safer than downloading and uploading the statefile with aws-cli.

When testing this with `terraform plan` instead of `terraform apply` against
staging and production (it's not possible to use this in dev or CI) the
following error was produced on the very first run:

    # check password store and if varables are accesible
    Initialized blank state with remote state enabled!
    Remote state configured and pulled.
    Remote state cache has no remote info
    make: *** [manually_upload_certs] Error 1

After which a `remote` hash is written to the remote statefile and
subsequent runs work as expected.