Add ACTION variable to manual Terraform commands
So that we can run `plan` instead of `apply` to safely review changes. I
made a change similar to this to my local working copy when reviewing a
[credentials PR](https://github.gds/government-paas/credentials/pull/33)
because I didn't want to make any unexpected changes to the live config.

I thought about setting a safe default action of `plan`, but decided against
it because people might assume that it had run `apply` and changes wouldn't
have been applied. It would have also been possible to re-use the `ACTION`
variable within the scripts but I thought it was better to explicitly parse
a CLI argument.

I've updated the documentation accordingly, listing both commands that you
are likely to use.