Add a new healthcheck-deployed resource
In order to not waste time or cause failures on a first run of the
pipeline we add a new piece of state, `healthcheck-deployed`, which can
be the strings "yes", "no", and defaults to no.

This is implemented in terms of the s3-iam type.

We update this flag in the autodelete and destroy pipelines (to set it
to "no" as we are destroying).

We consult this flag in the `predeploy/wait-for-availability-test` and
`availability-tests/run-tests` jobs, to skip these tests when the
healthcheck application is not yet deployed.