Add shellcheck to pipecleaner
Here we add a new class of check to pipecleaner `shellcheck`

When enabled (and it is enabled by default) when linting we look for
tasks where the run property looks like it is an inline shell script,
and invoke `shellcheck` over it to see if the script looks valid.

When an issue is found, it will be reported like so:

    ==  spec/fixtures/pipecleaner_shellcheck.yml  ==
    WARNING * Shellcheck: job='shellcheck', task='bad-compare', ~='
    In - line 2:
    if [ "red" == "blue" ]; then
               ^-- SC2039: In POSIX sh, == in place of = is undefined.
               ^-- SC2050: This expression is constant. Did you forget the $ on a variable?

    '

We also add a ginkgo test for pipecleaner.py, with a test fixture of a
pipeline with an error we know shellcheck will pick up on.

There's some minor refactoring of pipecleaner here to reduce having to
explictly add error classes.