Only deploy the graphite nozzle if missing/changed
Deploying the nozzle on every run means we lose stats from the firehose
while deployments are in progress.

This deploys the graphite-nozzle with the current nozzle git shorthash
in the name. If the current version is missing from the cf deployment it
is pushed, and all older versions are removed.