Graphite -> t2.medium
Graphite currently runs on m3.large. This VM has 8GB of memory and 2 CPUs.
Graphite is using 7% of CPU on average. It does not need such large VM.
T2.medium should be more than sufficient for it. Maybe even t2.small, but
that has only 2G of ram (might impact disk caching) and only 1 CPU & half
credits, which might not be enough if we put more load on graphite in the
future.