Add cloud_controller.admin_read_only to cf client
In cf v240 it has been added support for a read_only cloudcontroller
scope, see [1] and [2]. The scope has been added to the example
manifest.

I am not sure if the scope is required for the client providing that
it already has cloud_controller.admin, but as it has been added to the
upstream reference manifest, we add it too.

UAA will update the client during deployment.

[1] https://github.com/cloudfoundry/cf-release/commit/d9e11dd478d939bc2a135ee887bf35d32d82cca4
[2] https://www.pivotaltracker.com/n/projects/966314/stories/117558887