Add routing.router_groups.write to cf client
In order to allow the operator to change the port range for a router
group, we must asign routing.router_groups.write. This scope was
added in the example manifest of cf-release in [1] via [2]

Currently we don't enable the router api, but we want to keep
the configuration in sync with the upstream config.

UAA will update the clients during deployment.

[1] https://github.com/cloudfoundry/cf-release/commit/a332fb5b360297ec3c6251d5b5b01ff82b5f701d
[2] https://www.pivotaltracker.com/n/projects/1358110/stories/115980665