Add datadog agent to runtime config
We use our fork of the release as the original has some bugs. By adding
that job to the runtime config (BOSH addons) it is automatically
deployed to all servers managed by BOSH.

The datadog api key is read from pass and passed to the manifest as
environment variable by the pipeline. Similarly for the app key that we
will use in terraform.

This commit refers to custom_005 release which was used for review of
the PR for this branch (#481). It will be bumped to final release version
in later commits, as we don't include URL and sha1 hash here.