BOSH: remove ntp property
We are setting both AWS CPI and BOSH's ntp property to the same
values as default. There's no need to do that. We should only set
ntp property if we intend to use st. different than defauls. Moreover,
from the spec files:
> pool.ntp.org attempts to return IPs closest to your location
So the default is quite sane already. Let's keep it.