Reduce maximum DEPLOY_ENV string length
In 88f7e59 the maximum length of the DEPLOY_ENV variable was declared
as being 15 characters, however the derived name of the
aws_elb.logsearch_es-master enforces a hard limit of 32 characters,
leading to an error when terraforming the environment.

~~~
 * aws_elb.logsearch_es_master: "name" cannot be longer than 32 characters: "richardc-sept-logsearch-es-master"
~~~

Working backwards from this limit, this actually leaves us 12
characters to work with when naming our development environments.

~~~
irb(main):001:0> "richardc-sept-logsearch-es-master".size
=> 33
irb(main):002:0> "richardc-sept".size
=> 13
~~~