Create router monitor via terraform
We create a minimal datadog alert via terraform. It detects if a router
is missing by alerting if 10% data is missing in the cluster. As the
cluster is currently made of 2 members, if one is missing we would miss
50% data.

Terraform requires datadog api and app keys. They are read from pass and
passed to Terraform as TF_VARs by the pipeline.