Unify datadog tag configuration
Pull datadog tags into shared and inject them into manifests from there.
Because we reference meta.environment in them, we had to add this to
concourse and a dummy value to runtime-config.

Also, in order to be able to inject in cf-manifest, we merge in shared
datadog.yml, which also contains datadog-agent release. This means that
we don't have to upload the release explicitly, but can instead depend
on bosh fetching it for us, because it is specified in cf-manifest. We
use the same approach for collectd addon.