add enable/disable switch for datadog
In order to optimize costs, we don't want to send datadog metrics for dev,
unless we specifically decide to do so (e.g. for development of alerts).

Add a mechanism that selectively:
* runs datadog TF creation and destroy. This implies datadog terraform being
pulled out of CF terraform.
* configures real or bogus ('disabled') key for the agents. In case of
disabled key agents fail to submit metrics, meaning we don't get billed.

We still do install agents on all the VMs. We have a story to improve this
by analysing other mechanism we could use to disable datadog deployment:

https://www.pivotaltracker.com/n/projects/1275640/stories/130182525