Work around race condition in sync_admin_users
The script gets a list of users and determine which ones it should
remove. When it tries to remove the user, it may have already been
deleted by another process. This happened at least once with the
temporary user created by the continuous-smoke-tests. As the code raised
an exception, this would break the build.

It is safe to ignore as we would have deleted the user anyway. Now we
only log a message and continue.