Specify concourse manifest properties per-template
This is necessary to support the upcoming datadog work. The datadog and
concourse bosh releases both have a tag property, but they're
incompatable (one's an array, and the other a hash etc). To avoid this
conflict the properties need to be specified per template.

Having properties at this level are a feature of the new Bosh 2.0
manifest schema. bosh-init has some limited support for this:

- it treats `instance_groups` as an alias for `jobs`, and within each
  job it treats `jobs` as an alias for `templates`. All the other keys
  within these are as per the 1.0 manifest schema.
- It supports specifying properties at the template(job) level.

Because of these limitations, this retains the 1.0 schema names for the
groups to avoid confusion (where this would look like a 2.0 manifest, but
behave like a 1.0 manifest), and moves all the properties to be
per-template.

It was also necessary to retain the job-level properties key as an empty
hash to prevent bosh-init erroring when merging the properties.