datadog TF: set 2 mins explicitly
For some reason, even though TF docs report this as default for 'service'
type of monitors, when we edited monitors by hand and then re-run TF, we
noticed TF setting this back to 0. Also, alerts created by TF didn't seem
to work. We decided to set this explicitly in TF. The alerts now work.