Use our latest datadog-agent release
Use custom_006 release (after merging
https://github.com/alphagov/paas-datadog-agent-boshrelease/pull/1).

Compared to custom_005 we don't have binary changes,
but do this anyway since we always want to use the latest
version of our own bosh release dependencies.

Compared to usual way of squashing this into the 1st commit that
referenced the release we leave this as a separate commit because
we have changed the way of specifying the releases after the
custom_005 release was referenced for the 1st time.