Add keepalive on SSH connections
From `man ssh_config`:
Sets a timeout interval in seconds after which if no data has been
received from the server, ssh(1) will send a message through the
encrypted channel to request a response from the server.

We set it to 60s. Note that it will maintain the connection from
localhost to the jumpbox. The connection after that might be down.

Incidentally, if the jumpbox is down, after 3 failed keepalives (equal
to ServerAliveCountMax which defaults to 3) the connection will be
closed.