Stream the gorouter access logs to syslog
We want to get all the logs from the router in our logsearch platform.

Currently gorouter is configured to send the logs to syslog, but it is
not configured to stream the access logs.

By enabling this option, all the access_log  will be sent to syslog,
and from there to loggregator.

Additional metron load:

To ensure the metron agent on the router vms is happy with the
additional load (there were comments about potential increase in cpu) we
put it under stress in dev environment and monitored the cpu usage and
rate of dropped logs before and after the change We used vegetta from
two hosts to add some significant load to the healthcheck app.

We then checked the CPU usage of metron on the router VMs before and
after the change.

Before change:
	total vm cpu usage: ~8% / per cpu
	metron cpu: ~5%
	gorouter: ~7%
After change:
	total vm cpu usage: ~9& / per cpu
	metron cpu: ~6%
	gorouter: ~8%

so no significant change to report.