Self update pipelines before init
When we run self update pipelines in the init job, even though that job
might get updated by itself, it will not run again. This means that if
there are any new resources to bootstrap in this job, they would not
get bootstrapped and the remainder of the pipeline can get stuck waiting
on these forever.

Move self update to lock. Self update after you have acquired lock and
before triggering remaining jobs. This way if init job gets updated it
will run in its latest instance and bootstrap all the things it really
needs to.