CF update: max in flight 1
When we had 2, that meant that 2 cells both in the same zone would get
updated at the same time. In our current prod that means loss of 50% of
capacity - which is too much. We can bump max-in-flight per job if we
have some job in the future with lots of VMs and we want to update more
than one concurrently.