Concourse 2.2.1 upgrade
This release fixes the problem with disk space that we had a lot of
trouble with:

"BOSH-deployed workers will now be named after a frankenguid taking
parts from the BOSH instance ID and their hostname. This is to make it
so you can correlate the worker to the BOSH instance, while also
guaranteeing that when the worker is recreated it comes back under a new
name."

"the issue is that it comes back with the same name every time
to concourse, a worker coming back with the same name means all the
volumes that were previously on it should still be there
so it probably gets confused when it sees the imported volume for it in
the db and then it's not there"

As we use bosh-init to deploy Concourse the name was always "unknown".

Release notes promises big improvements to resources candidate selection
algorithm that should bring CPU usage down.

http://concourse.ci/downloads.html#v221
http://concourse.ci/downloads.html#v220