Define default value for datadog api key
Dadatog.yml is always included in the cf-manifest. This is because we
always define tags and other datadog properties and we grab from shared
datadog.yml file. In case we don't want to enable datadog and the api
key is missing for the environment, spruce rendering would fail as
there was no value defined for the api key.

Define default for this value.