Add cell memory use dashboard
We currently have no way of passing the number of zones into the manifest at
build time so the scale factor representing the number of zones after a failure
(2) is currently hardcoded in this template as "0.6666" for the AZ failure
graph. We need to revisit this if we introduce another zone.

This assumes that all cells are equally sized.