Run custom acceptance tests in parallel
Now that we have added the Ginkgo binary to the container used for
custom acceptance tests[1] we can use it to run tests in parallel.

We have added two extra tests which create service instances, so without
this parallelisation the tests would take 30-40 minutes. This change has
resulted in the tests taking 20-25 minutes.

The timeout of 30 minutes has not been preserved. Ginkgo does not provide
the same command line flag and would require specific timeouts to be added
to specs within the source code. While this is possible a lot of the tests
already have timeouts set. You can set a timeout on the Concourse task,
but this could result in the container being destroyed while there are still
RDS instances, therefore leaving orphaned resources in AWS.

[1] https://github.com/alphagov/paas-docker-cloudfoundry-tools/pull/71