Remove unused smoke-tests-log output and input
We used to send smoke tests log in the attachement of the emails, but
there were some issues with weird characters - see
26ed1169291acc08fbfc193f0a4925b2e1c19539

Ever since then the logs were not used by anything. Pipecleaner warns
about that. Remove the smoke-tests-log completely, so we get no errors
or warnings. In case we want to start sending logs again, we can
revert this commit.