Use meta.environment for concourse deployment name
We are using concourse riemann reporting function to get metrics
from concourse, as per the PR #511.

But we want to tag the metrics from concourse with a specific environment.
We will extend concourse release to add additional custom tags, but
as currently concourse sends the deployment name as a tag[1], we will use
it.

Additionally, in the CF deployment we also use the environment name as
deployment name[2], so this commit will add homogenity.

The change of the deployment name does not have any other impact, as
we use bosh-init that ignores it.

[1] https://github.com/concourse/concourse/blob/master/jobs/atc/templates/atc_ctl.erb#L157
[2] https://github.com/alphagov/paas-cf/blob/master/manifests/cf-manifest/manifest/000-base-cf-deployment.yml#L1-L5