Remove null router variables
For some reason, these variables were present but not set. When going
through spruce they get set to 'null'. This means that default value
is being used for these properties. And this also means that there is
no reason to include these properties in the manifest if we don't
intend to set them to any value.