Install consul_agent in all VMs
Install consul_agent in all VMs together with metron_agent, so that
metron_agent can resolve etcd.service.cf.internal.

We have setup metron_agent in all VMs as a addon in #495.

Metron agent requires connect to etcd to discover the dopplers, and it is
currently setup to discover the etcd by using `etcd.service.cf.internal`
that is served by consul.

We did not setup the consul_agent in all the VMs, so the ones missing it
had metron_agent installed, but unable to resolv etcd and, in
consequence, discover the dopplers. Because that, the metron_agents
were logging constantly the following error:

   {
      "file" : "/var/vcap/data/compile/metron_agent/loggregator/src/doppler/dopplerservice/finder.go",
      "timestamp" : 1475669685.3659,
      "source" : "metron",
      "line" : 139,
      "log_level" : "error",
      "message" : "Finder: Watch sent an error",
      "method" : "doppler/dopplerservice.(*Finder).run",
      "data" : {
	 "error" : "store request timed out"
      },
      "process_id" : 9989
   }

We will do some work in the future to reconsider this solution.
Alternative approaches are:

 * Configure etcd using static IPs
 * Do not install metron_agent just to configure rsyslog.