Add more disk&memory to ES vm types
As we are almost out of disk space on prod for ES machines we need to
increase it. The safe size here is double than previous one.

Also as the volume of logs increases and we do more searches and
indexing we need more memory and more CPUs to cope with the load.