Add script to check tags of a given git repo
This script can be used to check that the expected tagged version of
boshrelease is checked out before pushing.

We will use this feature to workaround the limitations on the
concourse git-resource when using tag filters: Regardless the tag-filter
the versions previously reported by the git-resource are still vaid
versions, so concourse might checkout the wrong version if no newer
version has been reported.