Upgrade datadog agent and enable bosh tags
The host names of our virtual machines are Bosh UUIDs in Datadog, which
is not very user friendly and makes metrics hard to find and read.

This PR was merged upstream [1] which adds tags to identify
the bosh job and index, which we can use in our checks and graphs. This
PR is merged back to our fork [2].

We add these changes to the manifest.

[1] https://github.com/onemedical/datadog-agent-boshrelease/pull/8
[2] https://github.com/alphagov/paas-datadog-agent-boshrelease/pull/7