Set datadog tags bosh-{az,job} for bosh&concourse
Bosh and Concourse are deployed using bosh-init. But bosh-init has
the limitation of setting the variable `spec.az` as `unknown` and
not seting `spec.job`.

Because that, these two machines have the wrong tags.

This code fixes this by overriding the bosh-job and bosh-az tags in the manifest.

In the case of bosh, as it can be deployed in different AZ, we get the
correct label from terraform, parametrised by the same variable.

In the case of concourse the AZ is hardcoded in the manifest.

We add tests to ensure that it is test. In the case of concourse, the
test will verify the value of the AZ set in the resource pool, to ensure
that one will be changed when the other is changed.