Set concourse login timeout to 5m on prod
This is to help us differentiate environments better. You should not need to
be logged in to production concourse for extended periods of time. Log in,
do the operation you need and close the session. This way you avoid
accidentally doing operations on prod instead of other environments, as our
concourses look identically now. Use 'prod' team to log in.