Improve space usage of concourse cleaner chrome plugin.
At present, there's some unnecessary padding at the top of the pipeline
display causing the scrollbars to appear etc.

This cleans thing up by allowing the main content block to use the full
space. As part of this, the enviuronment label has been moved to the
bottom, and floated on top to avoid it conflicting with the actual
pipeline elements - there's generally more room at the bottom of the
screen due to how the pipelines are layed out.